moved magicCard to model, and created cardoverView
