+++ MTGClient/src/mtgclient/model/MagicCard.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package mtg;
+package mtgclient.model;
 
 import java.util.ArrayList;
 import java.util.List;
